move contract method form inside ABI folder
